refactor(cli): use @svelte-vitals/core's relocated component-facts parser
Also fixes packages/cli/test/parse-file.test.ts's attrValueOf import,
which the task brief's file list omitted but has the same dependency
on the relocated svelte-ast helper.
