+++ packages/cli/src/providers/source/adapters/svelte-seo.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type { ImportInfo } from '../imports.js';
 import type { ComponentUse, ParsedTag } from '../parse.js';
-import { attrValueOf, attrTextOf } from '../parse.js';
+import { attrValueOf, attrTextOf } from '@svelte-vitals/core';
 import type { Adapter, AdapterResult } from './types.js';
 
 type Node = any;
